What Is SEO?
So you all know how important traffic is to a blog. Blogs that don’t have traffic are more than enough. Because everyone who blogs is expecting readers to come to their blog, but most blogs don’t get as much traffic. As such, foreigners alike are getting away from blogging.
So the main purpose of this SEO is to increase the traffic to your blog. Increase traffic means search engine word or word search through a search engine or a keyword to increase traffic to your blog. This traffic is also known as Organic Traffic.
more informion https://seomeaning.net/seo-meaning-the-word-seo-is-a-combination-of-the-first-three-letters-of-search-engine-optimization/